+++ Test++/nitf_test_create_nitf++.cpp
@@ -4,32 +4,27 @@
 
 #include "nitf_Test.h"
 
-struct nitf_test_create_nitf__ : public ::testing::Test {
+CODA_OSS_disable_warning_push
+#if _MSC_VER
+#pragma warning(disable: 4464) // relative include path contains '..'
+#endif
+#include <nitf/../../unittests/nitro_image_.c_>
+CODA_OSS_disable_warning_pop
+
+using namespace Microsoft::VisualStudio::CppUnitTestFramework;
+
+TEST_CLASS(nitf_test_create_nitf__) {
+public:
 	nitf_test_create_nitf__() {
 		// initialization code here
-		//const std::string NITF_PLUGIN_PATH = R"(C:\Users\jdsmith\source\repos\nitro\x64\Debug\share\nitf\plugins)";
 		sys::OS().setEnv("NITF_PLUGIN_PATH", nitf::Test::buildPluginsDir(), true /*overwrite*/);
 	}
 
-	void SetUp() {
-		// code here will execute just before the test ensues 
-	}
-
-	void TearDown() {
-		// code here will be called just after the test completes
-		// ok to through exceptions from here if need be
-	}
-
-	~nitf_test_create_nitf__() {
-		// cleanup any pending stuff, but no exceptions allowed
-	}
-
+	~nitf_test_create_nitf__() = default;
 	nitf_test_create_nitf__(const nitf_test_create_nitf__&) = delete;
 	nitf_test_create_nitf__& operator=(const nitf_test_create_nitf__&) = delete;
 
+	#define TEST_CASE(X) TEST_METHOD(X)
+	#include "nitf/unittests/test_create_nitf++.cpp"
 
-	// put in any custom data members that you need 
 };
-
-#define TEST_CASE(X) TEST_F(nitf_test_create_nitf__, X)
-#include "nitf/unittests/test_create_nitf++.cpp"
