+++ Test++/nitf_test_field++.cpp
@@ -2,5 +2,12 @@
 
 #include "nitf_Test.h"
 
-#define TEST_CASE(X) TEST(nitf_test_field__, X)
+using namespace Microsoft::VisualStudio::CppUnitTestFramework;
+
+TEST_CLASS(nitf_test_field__) {
+public:
+
+#define TEST_CASE(X) TEST_METHOD(X)
 #include "nitf/unittests/test_field++.cpp"
+
+};
