+++ Test++/nitf_test_image_io.cpp
@@ -1,10 +1,14 @@
 #include "pch.h"
 
-#include <windows.h>
-#include <combaseapi.h>
-#undef interface
-
 #include "nitf_Test.h"
 
-#define TEST_CASE(X) TEST(test_image_loading__, X)
+using namespace Microsoft::VisualStudio::CppUnitTestFramework;
+
+TEST_CLASS(test_image_loading__) {
+public:
+
+#define TEST_CASE(X) TEST_METHOD(X)
 #include "nitf/unittests/test_image_loading++.cpp"
+
+};
+std::string test_image_loading__::argv0;
