+++ Test++/nitf_test_image_segment_blank_nm_compression.cpp
@@ -2,5 +2,12 @@
 
 #include "nitf_Test.h"
 
-#define TEST_CASE(X) TEST(test_image_segment_blank_nm_compression, X)
+using namespace Microsoft::VisualStudio::CppUnitTestFramework;
+
+TEST_CLASS(test_image_segment_blank_nm_compression) {
+public:
+
+#define TEST_CASE(X) TEST_METHOD(X)
 #include "nitf/unittests/test_image_segment_blank_nm_compression.cpp"
+
+};
