+++ Test++/nitf_test_image_segment_computer.cpp
@@ -2,5 +2,12 @@
 
 #include "nitf_Test.h"
 
-#define TEST_CASE(X) TEST(test_image_segment_computer, X)
+using namespace Microsoft::VisualStudio::CppUnitTestFramework;
+
+TEST_CLASS(test_image_segment_computer) {
+public:
+
+#define TEST_CASE(X) TEST_METHOD(X)
 #include "nitf/unittests/test_image_segment_computer.cpp"
+
+};
