+++ Test++/nitf_test_j2k_read_tile.cpp
@@ -1,10 +1,13 @@
 #include "pch.h"
 
-#include <windows.h>
-#include <combaseapi.h>
-#undef interface
-
 #include "nitf_Test.h"
 
-#define TEST_CASE(X) TEST(test_j2k_read_tile, X)
+using namespace Microsoft::VisualStudio::CppUnitTestFramework;
+
+TEST_CLASS(test_j2k_read_tile) {
+public:
+
+#define TEST_CASE(X) TEST_METHOD(X)
 #include "nitf/unittests/test_j2k_read_tile.cpp"
+
+};
