+++ Test++/nitf_test_tre_create++.cpp
@@ -4,32 +4,19 @@
 
 #include "nitf_Test.h"
 
-struct nitf_test_tre_create__ : public ::testing::Test {
+using namespace Microsoft::VisualStudio::CppUnitTestFramework;
+
+TEST_CLASS(nitf_test_tre_create__) {
+public:
 	nitf_test_tre_create__() {
 		// initialization code here
-		//const std::string NITF_PLUGIN_PATH = R"(C:\Users\jdsmith\source\repos\nitro\x64\Debug\share\nitf\plugins)";
 		sys::OS().setEnv("NITF_PLUGIN_PATH", nitf::Test::buildPluginsDir(), true /*overwrite*/);
 	}
-
-	void SetUp() {
-		// code here will execute just before the test ensues 
-	}
-
-	void TearDown() {
-		// code here will be called just after the test completes
-		// ok to through exceptions from here if need be
-	}
-
-	~nitf_test_tre_create__() {
-		// cleanup any pending stuff, but no exceptions allowed
-	}
-
-
+	~nitf_test_tre_create__() = default;
 	nitf_test_tre_create__(const nitf_test_tre_create__&) = delete;
 	nitf_test_tre_create__& operator=(const nitf_test_tre_create__&) = delete;
 
-	// put in any custom data members that you need 
-};
-
-#define TEST_CASE(X) TEST_F(nitf_test_tre_create__, X)
+#define TEST_CASE(X) TEST_METHOD(X)
 #include "nitf/unittests/test_tre_create++.cpp"
+
+};
