+++ Test++/nitf_test_tre_read.cpp
@@ -4,32 +4,20 @@
 
 #include "nitf_Test.h"
 
-struct test_tre_read : public ::testing::Test {
+using namespace Microsoft::VisualStudio::CppUnitTestFramework;
+
+TEST_CLASS(test_tre_read) {
+public:
 	test_tre_read() {
 		// initialization code here
-		//const std::string NITF_PLUGIN_PATH = R"(C:\Users\jdsmith\source\repos\nitro\x64\Debug\share\nitf\plugins)";
 		sys::OS().setEnv("NITF_PLUGIN_PATH", nitf::Test::buildPluginsDir(), true /*overwrite*/);
 	}
-
-	void SetUp() {
-		// code here will execute just before the test ensues 
-	}
-
-	void TearDown() {
-		// code here will be called just after the test completes
-		// ok to through exceptions from here if need be
-	}
-
-	~test_tre_read() {
-		// cleanup any pending stuff, but no exceptions allowed
-	}
-
+	~test_tre_read() = default;
 	test_tre_read(const test_tre_read&) = delete;
 	test_tre_read& operator=(const test_tre_read&) = delete;
 
-
-	// put in any custom data members that you need 
-};
-
-#define TEST_CASE(X) TEST_F(test_tre_read, X)
+#define TEST_CASE(X) TEST_METHOD(X)
 #include "nitf/unittests/test_tre_read.cpp"
+
+};
+const char* test_tre_read::argv0 = nullptr;
