+++ Test++/pch.h
@@ -1,7 +1,11 @@
-//
-// pch.h
-// Header for standard system include files.
-//
+// pch.h: This is a precompiled header file.
+// Files listed below are compiled only once, improving build performance for future builds.
+// This also affects IntelliSense performance, including code completion and many code browsing features.
+// However, files listed here are ALL re-compiled if any one of them is updated between builds.
+// Do not add files here that you will be updating frequently as this negates the performance advantage.
+
+#ifndef PCH_H
+#define PCH_H
 #pragma once
 
 #pragma warning(disable: 4820) // '...': '...' bytes padding added after data member '...'
@@ -25,6 +29,8 @@
 #pragma warning(push)
 #pragma warning(disable: 4464) // relative include path contains '..'
 #include <nitf/coda-oss.hpp>
+#include <math/Round.h>
+#include <io/TempFile.h>
 #pragma warning(disable: 5031) // #pragma warning(pop): likely mismatch, popping warning state pushed in different file
 #pragma comment(lib, "io-c++")
 #pragma comment(lib, "except-c++")
@@ -33,20 +39,23 @@
 #pragma warning(pop)
 
 #pragma warning(push)
-#pragma warning(disable: 4388) // '...': signed / unsigned mismatch
-#pragma warning(disable: 4389) // '...': signed / unsigned mismatch
-#pragma warning(disable: 4800) // Implicit conversion from '...' to bool. Possible information loss
-#pragma warning(disable: 4625) // '...': copy constructor was implicitly defined as deleted
-#pragma warning(disable: 4626) // '...': assignment operator was implicitly defined as deleted
-#pragma warning(disable: 5026) // '...': move constructor was implicitly defined as deleted
-#pragma warning(disable: 5027) //	'...': move assignment operator was implicitly defined as deleted
-#include "gtest/gtest.h"
+#include "CppUnitTest.h"
 #pragma warning(pop)
 
-#include <import/nrt.h>
-#include <nitf/System.hpp>
+#include <import/nitf.h>
+
+#include <import/nitf.hpp>
+#include <nitf/TRE.hpp>
+#include <nitf/exports.hpp>
+#include <nitf/TREField.hpp>
+#include <nitf/FieldDescriptor.hpp>
+#include <nitf/TestingTest.hpp>
+#include <nitf/J2KComponent.hpp>
+#include <nitf/J2KContainer.hpp>
+#include <nitf/J2KReader.hpp>
+#include <nitf/J2KWriter.hpp>
 
 #include "nitf_Test.h"
 #include "Test.h"
 
-
+#endif //PCH_H
