--- UnitTest++/TestCase.h
@@ -1,2 +1,0 @@
-#pragma once
-#include "Test.h"
