--- UnitTest++/nitf_test_hash_table_1++.cpp
@@ -1,13 +1,0 @@
-#include "pch.h"
-
-#include "nitf_Test.h"
-
-using namespace Microsoft::VisualStudio::CppUnitTestFramework;
-
-TEST_CLASS(nitf_test_hash_table_1__) {
-public:
-
-#define TEST_CASE(X) TEST_METHOD(X)
-#include "nitf/unittests/test_hash_table_1++.cpp"
-
-};
