--- UnitTest++/nitf_test_image_segment_computer.cpp
@@ -1,13 +1,0 @@
-#include "pch.h"
-
-#include "nitf_Test.h"
-
-using namespace Microsoft::VisualStudio::CppUnitTestFramework;
-
-TEST_CLASS(test_image_segment_computer) {
-public:
-
-#define TEST_CASE(X) TEST_METHOD(X)
-#include "nitf/unittests/test_image_segment_computer.cpp"
-
-};
