--- UnitTest++/nitf_test_j2k_read_tile.cpp
@@ -1,13 +1,0 @@
-#include "pch.h"
-
-#include "nitf_Test.h"
-
-using namespace Microsoft::VisualStudio::CppUnitTestFramework;
-
-TEST_CLASS(test_j2k_read_tile) {
-public:
-
-#define TEST_CASE(X) TEST_METHOD(X)
-#include "nitf/unittests/test_j2k_read_tile.cpp"
-
-};
