--- UnitTest++/nitf_test_nitf_buffer_list.cpp
@@ -1,13 +1,0 @@
-#include "pch.h"
-
-#include "nitf_Test.h"
-
-using namespace Microsoft::VisualStudio::CppUnitTestFramework;
-
-TEST_CLASS(test_nitf_buffer_list) {
-public:
-
-#define TEST_CASE(X) TEST_METHOD(X)
-#include "nitf/unittests/test_nitf_buffer_list.cpp"
-
-};
