--- UnitTest++/nrt_Test.h
@@ -1,6 +1,0 @@
-#pragma once
-
-#if defined(MS_CPP_UNITTESTFRAMEWORK)
-#define __TEST_H__ // #include "../nrt/unittests/Test.h"
-#endif
-#include "Test.h"
